+++ R/FoReco-package.R
@@ -1,4 +1,4 @@
-#' FoReco: point forecast reconciliation
+#' FoReco: forecast reconciliation
 #'
 #' An R package offering classical (bottom-up and top-down), and modern (optimal and heuristic combination)
 #' forecast reconciliation procedures for cross-sectional, temporal, and cross-temporal
@@ -7,7 +7,7 @@
 #' @author Tommaso Di Fonzo and Daniele Girolimetto, Department of Statistical Sciences, University of Padua (Italy).
 #'
 #' @details
-#' The \code{FoReco} package is designed for point forecast reconciliation, a
+#' The \code{FoReco} package is designed for forecast reconciliation, a
 #' post-forecasting process aimed to improve the accuracy of the base
 #' forecasts for a system of linearly constrained (e.g. hierarchical/grouped) time series.
 #' The main functions are:
@@ -25,12 +25,15 @@
 #' }
 #'
 #' @references
-#' Di Fonzo, T., and Girolimetto, D. (2021), Cross-temporal forecast reconciliation:
+#' Di Fonzo, T., and Girolimetto, D. (2023), Cross-temporal forecast reconciliation:
 #' Optimal combination method and heuristic alternatives, \emph{International Journal
-#' of Forecasting}, in press \doi{10.1016/j.ijforecast.2021.08.004}.
+#' of Forecasting}, 39(1), 39-57 \doi{10.1016/j.ijforecast.2021.08.004}.
 #'
 #' Di Fonzo, T., Girolimetto, D. (2022), Forecast combination based forecast reconciliation:
 #' insights and extensions, \emph{International Journal of Forecasting}, in press.
+#'
+#' Girolimetto, D., Athanasopoulos, G., Di Fonzo, T., and Hyndman, R. J. (2023),
+#' Cross-temporal Probabilistic Forecast Reconciliation, \doi{10.48550/arXiv.2303.17277}.
 #'
 #' @import mathjaxr
 #' @docType package