+++ R/iterec.R
@@ -92,9 +92,9 @@
 #' table when the expected marginal totals are known, \emph{The Annals of Mathematical
 #' Statistics}, 11, 4, 427–444.
 #'
-#' Di Fonzo, T., and Girolimetto, D. (2021), Cross-temporal forecast reconciliation:
+#' Di Fonzo, T., and Girolimetto, D. (2023), Cross-temporal forecast reconciliation:
 #' Optimal combination method and heuristic alternatives, \emph{International Journal
-#' of Forecasting}, in press.
+#' of Forecasting}, 39(1), 39-57.
 #'
 #' Johnston, R.J., Pattie, C.J. (1993), Entropy-Maximizing and the Iterative Proportional
 #' Fitting Procedure, \emph{The Professional Geographer}, 45, 3, 317–322.