+++ R/tcsrec.R
@@ -22,7 +22,7 @@
 #' [lowest_freq' ...  highest_freq']'.
 #' @param hts_comb,thf_comb Type of covariance matrix (respectively
 #' (\mjseqn{n \times n}) and (\mjseqn{(k^\ast + m) \times (k^\ast + m)})) to
-#' be used in the cross-sectional and temporal reconciliation, see more in
+#' be used in the cross-sectional and temporal reconciliation. More details in
 #' \code{comb} param of \code{\link[FoReco]{htsrec}()} and
 #' \code{\link[FoReco]{thfrec}()}.
 #' @param res (\mjseqn{n \times N(k^\ast + m)}) matrix containing the residuals
@@ -37,7 +37,7 @@
 #' Athanasopoulos (2019), otherwise it is calculated as simple average of
 #' all the involved projection matrices at step 2 of the procedure (see
 #' Di Fonzo and Girolimetto, 2020).
-#' @param ... any other options useful for \code{\link[FoReco]{htsrec}()} and
+#' @param ... Any other options useful for \code{\link[FoReco]{htsrec}()} and
 #' \code{\link[FoReco]{thfrec}()}, e.g. \code{m}, \code{C} (or \code{Ut} and
 #' \code{nb}), \code{nn} (for non negativity reconciliation only at first step),
 #' \code{mse}, \code{corpcor}, \code{type}, \code{sol}, \code{settings},
@@ -50,8 +50,8 @@
 #' and \code{hts_comb} equal to either \code{"shr"} or \code{"wls"}.
 #'
 #' \strong{Warning},
-#' the two-step heuristic reconciliation allows non negativity constraints only in the first step.
-#' This means that non-negativity is not guaranteed in the final reconciled values.
+#' the two-step heuristic reconciliation allows considering non negativity constraints only in
+#' the first step. This means that non-negativity is not guaranteed in the final reconciled values.
 #'
 #' @return
 #' The function returns a list with two elements:
@@ -59,9 +59,9 @@
 #' \item{\code{M}}{Matrix which transforms the uni-dimensional reconciled forecasts of step 1 (projection approach) .}
 #'
 #' @references
-#' Di Fonzo, T., and Girolimetto, D. (2021), Cross-temporal forecast reconciliation:
+#' Di Fonzo, T., and Girolimetto, D. (2023), Cross-temporal forecast reconciliation:
 #' Optimal combination method and heuristic alternatives, \emph{International Journal
-#' of Forecasting}, in press.
+#' of Forecasting}, 39(1), 39-57.
 #'
 #' Kourentzes, N., Athanasopoulos, G. (2019), Cross-temporal coherent forecasts
 #' for Australian tourism, \emph{Annals of Tourism Research}, 75, 393-409.